+++ src/GluttonyCombo/GluttonyCombo/Combos/CustomComboPreset.cs
@@ -1,4 +1,4 @@
-﻿#region
+#region
 
 using ECommons.ExcelServices;
 using GluttonyCombo.Attributes;
@@ -1623,18 +1623,6 @@ public enum Preset
 
     #region BLUE MAGE
 
-    [AutoAction(false, false)]
-    [ReplaceSkill(BLU.SonicBoom)]
-    [JobInfo(Job.BLU)]
-    [AdvancedCombo]
-    BLU_ST_AdvancedMode = 70030,
-
-    [AutoAction(false, true)]
-    [ReplaceSkill(BLU.SonicBoom)]
-    [JobInfo(Job.BLU)]
-    [AdvancedCombo]
-    BLU_Heal_AdvancedMode = 70031,
-
     [ReplaceSkill(BLU.MoonFlute)]
     [BlueInactive(BLU.Whistle, BLU.Tingle, BLU.RoseOfDestruction, BLU.MoonFlute, BLU.JKick, BLU.TripleTrident,
         BLU.Nightbloom, BLU.WingedReprobation, BLU.SeaShanty, BLU.BeingMortal, BLU.ShockStrike, BLU.Surpanakha,
@@ -1769,7 +1757,7 @@ public enum Preset
     [JobInfo(Job.BLU)]
     BLU_PeatClean = 70023,
 
-    // Last value = 70031
+    // Last value = 70025
 
     #endregion